Start migrating to golang to call the api
Golang will be more portable and easier to share with other parents. I
want to do it in small steps though, so I'm starting with keeping the
ruby script and having it shell out to golang for compass API calls.

Over time I can port more and more to go. Maybe one day we'll reach
100%?